requirements.txt
django==2.2
django-allauth
djangorestframework
django-cors-headers
django-rest-swagger
drf-extensions
django-filter
django-rest-auth
pillow
gunicorn
django-import-export
